Conflicts:
	autoMATE_client/autoMATEServer/test/com/automate/protocol/client/messages/ClientCommandMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/client/subParsers/ClientCommandMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerAuthenticationMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerCommandMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerNodeListMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerPingMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerStatusUpdateMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/messages/ServerWarningMessageTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerAuthenticationMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerCommandMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerNodeListMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerPingMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerStatusUpdateMessageSubParserTest.java
	autoMATE_client/autoMATEServer/test/com/automate/protocol/server/subParsers/ServerWarningMessageSubParserTest.java

+++ autoMATE_client/autoMATEServer/src/com/automate/protocol/client/subParsers/ClientStatusUpdateMessageSubParser.java
@@ -13,7 +13,7 @@ import com.automate.util.xml.XmlFormatException;
 
 public class ClientStatusUpdateMessageSubParser extends ClientMessageSubParser<ClientStatusUpdateMessage> {
 
-	private int nodeId;
+	private long nodeId;
 	
 	/* (non-Javadoc)
 	 * @see com.automate.protocol.client.subParsers.ClientMessageSubParser#parseXml(java.lang.String)
@@ -47,7 +47,7 @@ public class ClientStatusUpdateMessageSubParser extends ClientMessageSubParser<C
 			Attributes attributes) throws SAXException {
 		if(qName.equals("status-update")) {
 			try {				
-				nodeId = Integer.parseInt(attributes.getValue("node-id"));
+				nodeId = Long.parseLong(attributes.getValue("node-id"));
 			} catch(NumberFormatException e) {
 				throw new SAXException(e);
 			}
